+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+- Converted all code to use the second major revision of Go protocol buffer, a.k.a APIv2 ()
+
+- Added support for variables in `client.GraphQLQuery`.
+
 - Initial (work in progress) version of GraphQL over gRPC support.
 
-- Token management & persistent storage
+- Added support for unauthenticated endpoints.
+
+- Token management & persistent storage.